Add the missing "symfony/http-kernel" dependency
Although the driver uses Symfony\Component\HttpKernel\* , this is not mentioned in composer.json. Many projects have this dependency already, but for a new project, following the installation steps (`composer backup-manager/symfony "^1.0"`) is insufficient: you would end up with a Symfony driver that's missing a dependency. Explicitly adding it into composer.json fixes the issue.